Adds the ability to save/edit/delete an env var
